% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParaInfer.R
\name{ParaInfer}
\alias{ParaInfer}
\title{Inferring the model parameters}
\usage{
ParaInfer(data, ngener, nsample, acc_rate = 0.25)
}
\arguments{
\item{data}{The data is fitted to infer the model parameters.}

\item{ngener}{The number of the generations of the algorithm.}

\item{nsample}{The number of the samples for each generation.}

\item{acc_rate}{The acceptance rate that determines the best portion of the samples to be accepted.}
}
\value{
The abundance matrix for all plots and species through time.
}
\description{
Inferring the model parameters
}
\examples{
plantsim(nplot = 5, nspe = 2, t = 10, ini_abundance = matrix(1:10, 5, 2),
 growth_rate = 1, interaction_matrix = matrix(0.001, 2, 2), st_portion = 0.7, surv_rate = 1)
}
