% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlantSim.R
\name{plantsim}
\alias{plantsim}
\title{Simulate a plant community}
\usage{
plantsim(
  nplot,
  nspe,
  t,
  ini_abundance,
  growth_rate,
  interaction_matrix,
  st_portion,
  filesave = NULL,
  vary_k = FALSE,
  distribution = "uniform"
)
}
\arguments{
\item{nplot}{The number of the plots in the community}

\item{nspe}{The number of the species}

\item{t}{The total time steps}

\item{ini_abundance}{The initial abundance of the species}

\item{growth_rate}{The growth rate for each plot and each species. It could be universally same.}

\item{interaction_matrix}{The interaction matrix among species}

\item{st_portion}{The portion of the seedlings that stay at the parental plot.}

\item{filesave}{The file name to be saved.}

\item{vary_k}{Set vary_k be a vector of two entries providing the mean and the sd. Otherwise, it is FALSE giving the same k = 1 to all plots.}

\item{distribution}{Distribution mode.}
}
\value{
The abundance matrix for all plots and species through time.
}
\description{
Simulate a plant community
}
\examples{
plantsim(nplot = 5, nspe = 2, t = 10, ini_abundance = matrix(1:10, 5, 2),
 growth_rate = 1, interaction_matrix = matrix(0.001, 2, 2), st_portion = 0.7)
}
